Issue #15 - Fix hang on connection error
The application using this library hangs (never exits) if the SSH tunnel is being
used by a lots of goroutines and an error occurs in the forward() method when a connection
is being made. Connection attempts seem to intermittently fail, and this somehow leads to the code
never exiting when complete.
I found that the connection attempt would succeed after 1 or 2 retries, then the application would later
exit like normal. This PR adds up to 10 retries. I have been using this fix successfully in an ETL application
that blasts data into Postgres tables through up to 70 goroutines, and it has been rock solid.