Add connection retry with MaxConnectionAttempts (#16)
The application using this library hangs (never exits) if the SSH tunnel is being
used by a lots of goroutines and an error occurs in the forward() method when a connection
is being made. Connection attempts seem to intermittently fail, and this somehow leads to the code
never exiting when complete.

I found that the connection attempt would succeed after 1 or 2 retries, then the application would later
exit like normal. This PR adds a optional retry mechanism. It must be enabled with MaxConnectionAttempts.

Fixed #15